requirements.txt
fastapi
uvicorn
lark-parser
sympy
streamlit
graphviz
pytest
black
ruff
requests  
plotly
kaleido
python-dotenv
kaleido